requirements.txt
# Core dependencies
psycopg2-binary

# Email support
email-validator

# Configuration
python-dotenv
pyyaml

# Logging
python-json-logger

# Utilities
requests
pydantic
fastapi

# Testing & QA
pytest
pytest-asyncio
pytest-cov
flake8
black
isort
mypy

# Security & API Enhancements
passlib[bcrypt]
python-jose[cryptography]
python-multipart
slowapi

# Monitoring
psutil

annotated-doc
annotated-types
anyio
APScheduler
bcrypt
certifi
cffi
charset-normalizer
click
colorama
coverage
cryptography
Deprecated
dnspython
ecdsa
email-validator
fastapi
flake8
greenlet
h11
httpcore
httpx
idna
iniconfig
isort
librt
limits
mccabe
mypy
mypy_extensions
numpy
packaging
passlib
pathspec
platformdirs
pluggy
psutil
psycopg2-binary
pyasn1
pycodestyle
pycparser
pydantic
pydantic_core
pyflakes
Pygments
pytest
pytest-asyncio
pytest-cov
python-dateutil
python-dotenv
python-jose
python-json-logger
python-multipart
pytokens
pytz
PyYAML
redis
requests
rsa
six
slowapi
SQLAlchemy
starlette
typing-inspection
typing_extensions
tzdata
tzlocal
urllib3
uvicorn
wrapt
